@@ -4,7 +4,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,10 +27,15 @@ dependencies:
 
 dependency_overrides:
   path: 1.7.0
+  quiver: 2.0.3
+  rxdart: ^0.23.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  hive_generator: ^0.7.0+2
+  build_runner: ^1.10.0
 
 
 flutter:
