@@ -1,8 +1,7 @@
 import 'package:clean_architecture_template/core/exceptions/exceptions.dart';
-import 'package:clean_architecture_template/core/network/basic_request/basic_request.dart';
 import 'package:clean_architecture_template/core/network/basic_request/base_options.dart';
+import 'package:clean_architecture_template/core/network/basic_request/basic_request.dart';
 import 'package:dio/dio.dart';
-import 'package:flutter/foundation.dart';
 
 class BasicDioRequestImpl implements BasicRequest {
   Dio dio;
@@ -50,7 +49,8 @@ class BasicDioRequestImpl implements BasicRequest {
 
       return response;
     } on DioError catch (e) {
-      throw DioException(message: e.message);
+      print(e.message);
+      throw e;
     } catch (e) {
       throw e;
     }
