@@ -1,9 +1,10 @@
 import 'dart:async';
+import 'package:clean_architecture_template/core/clean_utils/entity/entity_base.dart';
+import 'package:clean_architecture_template/core/clean_utils/repository/repository_base.dart';
 import 'package:clean_architecture_template/core/database/data_object.dart';
 import 'package:clean_architecture_template/core/database/database.dart';
 import 'package:clean_architecture_template/core/exceptions/exceptions.dart';
 import 'package:clean_architecture_template/core/network/connected/connected.dart';
-import 'package:clean_architecture_template/core/repository/repository_base.dart';
 import 'package:dartz/dartz.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:meta/meta.dart';
@@ -23,10 +24,27 @@ class FeatureNameRepositoryImpl implements FeatureNameRepository {
       @required this.db});
 
   @override
-  Stream<Either<Exception, FeatureNameEntity>> getFeatureName(
-      dynamic properties) {
+  Stream<Either<Exception, EntityBase>> getFeatureName(dynamic properties) {
     return RepositoryBase(
-            db: db, remoteDataSource: remoteDataSource, connected: connected)
+            dbName: 'FeatureNameCache',
+            db: db,
+            remoteDataSource: remoteDataSource,
+            connected: connected)
         .getDataSource(properties);
   }
+
+  /*Future<Either<Exception, FeatureNameEntity>> connectToServer(String aid) async {
+    try {
+      final remoteSource = await remoteDataSource.getFeatureName(aid: aid);
+
+      saveData();
+
+      print(remoteSource.data);
+
+      return Right(remoteSource);
+    } catch (e){
+      return Left(e);
+    }
+  }*/
+
 }
